concurrent download all images

+++ downloader/download_images.py
@@ -1,7 +1,8 @@
 import json
 import os
 import requests
-import uuid
+from concurrent.futures import ThreadPoolExecutor, as_completed
+import time
 
 import click
 import cv2
@@ -36,9 +37,9 @@ def decode_image(image_data):
 @click.option('--output_path', required=True, help='The output path for saving the downloaded images')
 def main(metadata_file_path: str, output_path: str):
     image_info = list(map(json.loads, open(metadata_file_path)))
-    image_urls = [info['image_url'] for info in image_info]
-    image_names = [os.path.basename(info['image']) for info in image_info]
-    print(f'Number of images: {len(image_urls)}')
+    image_urls = [info['image_url'] for info in image_info if 'image_url' in info]
+    image_names = [os.path.basename(info['image']) for info in image_info if "image" in info]
+    print(f'{metadata_file_path} | Number of images: {len(image_urls)}')
 
     metadata_file_name = os.path.basename(metadata_file_path)
     split_pos = metadata_file_name.rfind('_')
@@ -48,16 +49,29 @@ def main(metadata_file_path: str, output_path: str):
     output_image_path = os.path.join(output_path, object_type, image_type)
     os.makedirs(output_image_path, exist_ok=True)
 
-    for url, name in tqdm.tqdm(zip(image_urls, image_names), total=len(image_urls)):
+    def process(url, name):
+        new_file_path = os.path.join(output_image_path, name)
+        if os.path.exists(new_file_path):
+            return
+
         try:
             response = requests.get(url, timeout=TIME_OUT, headers=HEADERS)
             image_data = response.content
             image = decode_image(image_data)
 
-            new_file_path = os.path.join(output_image_path, name)
             cv2.imwrite(new_file_path, image)
         except:
-            continue
+            pass
+
+    threads = []
+    with ThreadPoolExecutor(max_workers=24) as executor:
+        for url, name in tqdm.tqdm(zip(image_urls, image_names), total=len(image_urls)):
+            threads.append(
+                executor.submit(process, url, name))
+            time.sleep(0.01)
+
+        for task in tqdm.tqdm(as_completed(threads), total=len(image_urls), desc="Processing requests"):
+            task.result()
 
 
 if __name__ == '__main__':
